+++ tests/integration/test_bootstrap.py
@@ -12,7 +12,7 @@ def test_bootstrap(tmp_path, testdata):
         shutil.copy2(testdata / 'nonmem' / 'pheno.ext', tmp_path)
         shutil.copy2(testdata / 'nonmem' / 'pheno.lst', tmp_path)
         model = Model('pheno.mod')
-        model.dataset_path = tmp_path / 'pheno.dta'
+        model.datainfo.path = tmp_path / 'pheno.dta'
         model.modelfit_results.ofv  # Read in results
         res = Bootstrap(model, 3).run()
         assert len(res.parameter_estimates) == 3