+++ tests/integration/test_iiv.py
@@ -16,7 +16,7 @@ def test_iiv_block_structure(tmp_path, testdata, list_of_parameters, no_of_model
         shutil.copy2(testdata / 'nonmem' / 'models' / 'mox2.mod', tmp_path)
         shutil.copy2(testdata / 'nonmem' / 'models' / 'mx19B.csv', tmp_path)
         model_start = Model('mox2.mod')
-        model_start.dataset_path = tmp_path / 'mx19B.csv'
+        model_start.datainfo.path = tmp_path / 'mx19B.csv'
 
         add_peripheral_compartment(model_start)
         add_iiv(model_start, list_of_parameters, 'add')
@@ -36,7 +36,7 @@ def test_iiv_no_of_etas(tmp_path, testdata):
         shutil.copy2(testdata / 'nonmem' / 'models' / 'mox2.mod', tmp_path)
         shutil.copy2(testdata / 'nonmem' / 'models' / 'mx19B.csv', tmp_path)
         model_start = Model('mox2.mod')
-        model_start.dataset_path = tmp_path / 'mx19B.csv'
+        model_start.datainfo.path = tmp_path / 'mx19B.csv'
 
         res = run_tool('iiv', 'brute_force_no_of_etas', model=model_start)