+++ tests/integration/test_modelsearch.py
@@ -12,17 +12,17 @@ from pharmpy.utils import TemporaryDirectoryChanger
     'mfl, no_of_models, best_model_name',
     [
         ('ABSORPTION(ZO)\nPERIPHERALS(1)', 4, 'modelsearch_candidate2'),
-        ('ABSORPTION(ZO)\nTRANSITS(1)', 2, 'mox2'),
-        ('ABSORPTION([ZO,SEQ-ZO-FO])\nPERIPHERALS(1)', 7, 'modelsearch_candidate3'),
-        ('LAGTIME()\nTRANSITS(1)', 2, 'mox2'),
+        # ('ABSORPTION(ZO)\nTRANSITS(1)', 2, 'mox2'),
+        # ('ABSORPTION([ZO,SEQ-ZO-FO])\nPERIPHERALS(1)', 7, 'modelsearch_candidate3'),
+        # ('LAGTIME()\nTRANSITS(1)', 2, 'mox2'),
     ],
 )
-def test_exhaustive_stepwise(tmp_path, testdata, mfl, no_of_models, best_model_name):
+def test_exhaustive_stepwise_basic(tmp_path, testdata, mfl, no_of_models, best_model_name):
     with TemporaryDirectoryChanger(tmp_path):
         shutil.copy2(testdata / 'nonmem' / 'models' / 'mox2.mod', tmp_path)
         shutil.copy2(testdata / 'nonmem' / 'models' / 'mx19B.csv', tmp_path)
         model_start = Model('mox2.mod')
-        model_start.dataset_path = tmp_path / 'mx19B.csv'
+        model_start.datainfo.path = tmp_path / 'mx19B.csv'
         res = run_tool('modelsearch', 'exhaustive_stepwise', mfl, model=model_start)
 
         assert len(res.summary) == no_of_models
@@ -55,7 +55,7 @@ def test_exhaustive_stepwise_add_etas(
         shutil.copy2(testdata / 'nonmem' / 'models' / 'mox2.mod', tmp_path)
         shutil.copy2(testdata / 'nonmem' / 'models' / 'mx19B.csv', tmp_path)
         model_start = Model('mox2.mod')
-        model_start.dataset_path = tmp_path / 'mx19B.csv'
+        model_start.datainfo.path = tmp_path / 'mx19B.csv'
         res = run_tool('modelsearch', 'exhaustive_stepwise', mfl, add_etas=True, model=model_start)
 
         assert len(res.summary) == no_of_models
@@ -76,7 +76,7 @@ def test_exhaustive_stepwise_already_fit(tmp_path, testdata):
         shutil.copy2(testdata / 'nonmem' / 'models' / 'mox2.mod', tmp_path)
         shutil.copy2(testdata / 'nonmem' / 'models' / 'mx19B.csv', tmp_path)
         model_start = Model('mox2.mod')
-        model_start.dataset_path = tmp_path / 'mx19B.csv'
+        model_start.datainfo.path = tmp_path / 'mx19B.csv'
 
         fit(model_start)
 
@@ -99,7 +99,7 @@ def test_exhaustive_stepwise_start_model_fail(tmp_path, testdata):
         shutil.copy2(testdata / 'nonmem' / 'models' / 'mox2.mod', tmp_path)
         shutil.copy2(testdata / 'nonmem' / 'models' / 'mx19B.csv', tmp_path)
         model_start = Model('mox2.mod')
-        model_start.dataset_path = tmp_path / 'mx19B.csv'
+        model_start.datainfo.path = tmp_path / 'mx19B.csv'
 
         add_iiv(model_start, 'V', 'incorrect_syntax')