+++ tests/nonmem/test_nonmem_model.py
@@ -754,7 +754,7 @@ def test_cmt_warning(testdata):
 
     model_str = model_original.model_code.replace('CMT=DROP', 'CMT')
     model = Model(StringIO(model_str))
-    model.dataset_path = model_original.dataset_path
+    model.datainfo.path = model_original.datainfo.path
 
     set_zero_order_absorption(model)